Disable cleanup for failed integration tests

+++ internal/integration/testcase_test.go
@@ -69,7 +69,7 @@ func (t testcase) Init() testcase {
 	return t
 }
 
-func (t testcase) CreatePipeline(ctx context.Context) (string, func()) {
+func (t testcase) CreatePipeline(ctx context.Context) (string, func(testcase) func()) {
 	t.Helper()
 
 	tpl, err := template.ParseFS(fixtures, fmt.Sprintf("fixtures/%s", t.Fixture))
@@ -89,9 +89,11 @@ func (t testcase) CreatePipeline(ctx context.Context) (string, func()) {
 	})
 	require.NoError(t, err)
 
-	return *pipeline.GraphQLID, func() {
-		if !preservePipelines {
-			t.deletePipeline(ctx)
+	return *pipeline.GraphQLID, func(t testcase) func() {
+		return func() {
+			if !preservePipelines && !t.Failed() {
+				t.deletePipeline(ctx)
+			}
 		}
 	}
 }
